Refactor Flight Page and sub pages
