+++ frontend/src/App.tsx
@@ -7,8 +7,8 @@ import FlightPage from "./pages/FlightPage/FlightPage.tsx";
 import axios from 'axios';
 import {useEffect, useState} from "react";
 import {Flight} from "./types/model/dataType.ts";
-import AddFlightPage from "./pages/AddFlightPage/AddFlightPage.tsx";
-import FlightDetailPage from "./pages/FlightDetailPage/FlightDetailPage.tsx";
+import AddFlightPage from "./pages/FlightPage/AddFlightPage/AddFlightPage.tsx";
+import FlightDetailPage from "./pages/FlightPage/FlightDetailPage/FlightDetailPage.tsx";
 import {BasicUser, UserForLogin} from "./types/auth/userType.ts";
 import LoginPage from "./pages/LoginPage/LoginPage.tsx";
 import SignupPage from "./pages/LoginPage/SignupPage.tsx";
