+++ frontend/src/pages/FlightPage/FlightDetailPage/FlightDetailPage.css
